+++ decal.cpp
@@ -338,7 +338,7 @@ void Decal::PreRenderText()
    BITMAPINFO bmi = {};
    bmi.bmiHeader.biSize = sizeof(BITMAPINFOHEADER);
    bmi.bmiHeader.biWidth = m_textImg->width();
-   bmi.bmiHeader.biHeight = -m_textImg->height();
+   bmi.bmiHeader.biHeight = -(LONG)m_textImg->height();
    bmi.bmiHeader.biPlanes = 1;
    bmi.bmiHeader.biBitCount = 32;
    bmi.bmiHeader.biCompression = BI_RGB;
@@ -378,9 +378,9 @@ void Decal::PreRenderText()
    // Copy and set to opaque
    D3DCOLOR* __restrict bitsd = (D3DCOLOR*)bits;
    D3DCOLOR* __restrict dest = (D3DCOLOR*)m_textImg->data();
-   for (int i = 0; i < m_textImg->height(); i++)
+   for (unsigned int i = 0; i < m_textImg->height(); i++)
    {
-      for (int l = 0; l < m_textImg->width(); l++, dest++, bitsd++)
+      for (unsigned int l = 0; l < m_textImg->width(); l++, dest++, bitsd++)
          *dest = *bitsd | 0xFF000000u;
       dest += m_textImg->pitch()/4 - m_textImg->width();
    }
